pyproject.toml
[project]
name = "resumo-evolucoes-clinicas"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.0",
    "openai>=1.76.0",
    "playwright>=1.58.0",
    "python-dotenv>=1.2.2",
]

[dependency-groups]
dev = []
